# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/users:
    get:
      description: Requires users:view.
      operationId: UsersController_findAll_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a5
        - bearer: []
      summary: List users
      tags: &a6
        - Users
    post:
      description: Requires users:create.
      operationId: UsersController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
      security: *a5
      summary: Create a user and assign roles
      tags: *a6
  /v1/users/{id}:
    get:
      description: Requires users:view.
      operationId: UsersController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a5
      summary: Get a user
      tags: *a6
    patch:
      description: Requires users:edit. status=inactive revokes access immediately.
      operationId: UsersController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
      security: *a5
      summary: Edit / deactivate a user
      tags: *a6
  /v1/users/{id}/roles:
    put:
      description: Requires users:edit.
      operationId: UsersController_setRoles_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserRolesDto"
      responses:
        "200":
          description: ""
      security: *a5
      summary: Set a user’s roles
      tags: *a6
...
  /v1/sales:
    post:
      description: Requires sales:create. Server generates the Sale ID; item snapshots
        stay NULL.
      operationId: SalesController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleDto"
      responses:
        "201":
          description: ""
      security: &a23
        - bearer: []
      summary: Enter a sale
      tags: &a24
        - Sales & Validation
    get:
      description: Requires sales:view. Filters status/rep_id/client_id/date. Scoped
        per caller in the query.
      operationId: SalesController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - entered
              - validated
              - in_pay_run
              - paid
              - clawed_back
              - deleted
        - name: rep_id
          required: false
          in: query
          description: Filter by rep (intersected with the caller’s scope).
          schema:
            type: string
        - name: client_id
          required: false
          in: query
          schema:
            type: string
        - name: date_from
          required: false
          in: query
          description: sale_date >= (YYYY-MM-DD).
          schema:
            example: 2026-01-01
            type: string
        - name: date_to
          required: false
          in: query
          description: sale_date <= (YYYY-MM-DD).
          schema:
            example: 2026-01-31
            type: string
      responses:
        "200":
          description: ""
      security: *a23
      summary: List sales
      tags: *a24
  /v1/sales/bulk-validate:
    post:
      description: Requires sales:approve. NOT a file upload — client-report ingestion
        is the Import module.
      operationId: SalesController_bulkValidate_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkValidateDto"
      responses:
        "200":
          description: ""
      security: *a23
      summary: Batch-validate selected sales (queue bulk-select)
      tags: *a24
  /v1/sales/{id}:
    get:
      description: Requires sales:view. Includes items + derived pay period.
      operationId: SalesController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a23
      summary: Get a sale
      tags: *a24
    patch:
      description: Requires sales:edit. Only Entered sales; client/sale_date/mpu_id
        are immutable.
      operationId: SalesController_edit_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSaleDto"
      responses:
        "200":
          description: ""
      security: *a23
      summary: Edit an entered sale
      tags: *a24
    delete:
      description: Requires sales:delete. entered|validated → status=deleted (row
        preserved); rejected once paid.
      operationId: SalesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a23
      summary: Delete a pre-payout sale (soft)
      tags: *a24
  /v1/sales/{id}/validate:
    post:
      description: Requires sales:approve. Approval gate; never changes the pay
        period. Optional greenfield confirm.
      operationId: SalesController_validate_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateSaleDto"
      responses:
        "200":
          description: ""
      security: *a23
      summary: Validate a sale (entered → validated)
      tags: *a24
  /v1/sales/{id}/greenfield:
    post:
      description: Requires sales:approve. PROPOSED two-step; recomputes
        counts_toward_tally. Pay Run consumes at close.
      operationId: SalesController_setGreenfield_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetGreenfieldDto"
      responses:
        "200":
          description: ""
      security: *a23
      summary: Confirm/clear greenfield (PROPOSED — SRS §17)
      tags: *a24
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
          example: jane@redwave.local
        password:
          type: string
          writeOnly: true
          minLength: 8
        full_name:
          type: string
          example: Jane Doe
        phone:
          type: string
        avatar_url:
          type: string
        role_ids:
          description: Role ids to assign on creation.
          type: array
          items:
            type: string
      required:
        - email
        - password
        - full_name
    UpdateUserDto:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
        status:
          type: string
          enum:
            - active
            - inactive
          description: Set inactive to deactivate (immediate revoke).
    SetUserRolesDto:
      type: object
      properties:
        role_ids:
          description: The complete set of role ids the user should hold.
          type: array
          items:
            type: string
      required:
        - role_ids
...
    SaleItemInput:
      type: object
      properties:
        product_id:
          type: string
          description: A product belonging to the sale’s client (active).
      required:
        - product_id
    CreateSaleDto:
      type: object
      properties:
        client_id:
          type: string
          description: The client (must be an existing active client).
        rep_id:
          type: string
          description: Rep the sale belongs to. Defaults to the caller’s own rep;
            scope-authorized.
        sale_date:
          type: string
          example: 2026-01-10
          description: YYYY-MM-DD. Defaults to today. GOVERNS the pay period.
        activation_date:
          type: string
          example: 2026-01-15
          description: YYYY-MM-DD. Reference only — drives no logic.
        customer_name:
          type: string
        street:
          type: string
        city:
          type: string
        province_state:
          type: string
        postal_code:
          type: string
        mpu_id:
          type: string
          description: Client house ID, where supplied (some clients do not provide one).
        is_greenfield:
          type: boolean
          description: Rep’s greenfield REQUEST at entry; an admin confirms/clears at
            validation.
        items:
          description: One or more products on this sale.
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required:
        - client_id
        - customer_name
        - street
        - city
        - province_state
        - postal_code
        - items
    BulkValidateDto:
      type: object
      properties:
        sale_ids:
          description: Sale ids to validate. Non-entered sales are reported, not thrown.
          type: array
          items:
            type: string
      required:
        - sale_ids
    UpdateSaleDto:
      type: object
      properties:
        customer_name:
          type: string
        street:
          type: string
        city:
          type: string
        province_state:
          type: string
        postal_code:
          type: string
        activation_date:
          type: string
          example: 2026-01-15
          description: YYYY-MM-DD. Reference only.
    ValidateSaleDto:
      type: object
      properties:
        is_greenfield:
          type: boolean
          description: "Optional: confirm (true) or clear (false) greenfield while
            validating."
    SetGreenfieldDto:
      type: object
      properties:
        is_greenfield:
          type: boolean
          description: true = confirmed greenfield (internet excluded from tally); false =
            cleared.
      required:
        - is_greenfield
...